fix(frontend): move toaster to app to avoid issues
