+++ frontend/apps/open/src/components/providers.tsx
@@ -1,12 +1,13 @@
 "use client";
 
-import { Toaster } from "@vulkanlabs/base/ui";
+import { ThemeProvider } from "next-themes";
+import { Toaster } from "./toaster";
 
 export function Providers({ children }: { children: React.ReactNode }) {
     return (
-        <>
+        <ThemeProvider attribute="class" defaultTheme="light">
             {children}
             <Toaster />
-        </>
+        </ThemeProvider>
     );
 }
