+++ frontend/packages/base/src/ui/index.ts
@@ -20,7 +20,6 @@ export * from "../components/ui/select";
 export * from "../components/ui/separator";
 export * from "../components/ui/sheet";
 export * from "../components/ui/skeleton";
-export * from "../components/ui/sonner";
 export * from "../components/ui/switch";
 export * from "../components/ui/table";
 export * from "../components/ui/tabs";
